@@ -7,60 +7,60 @@
   "pnpm": {
     "overrides": {
       "axios": ">=1.7.4",
-      "d3-color": ">=3.1.0"
+      "d3-color": ">=3.1.0",
+      "cross-spawn": "^7.0.6"
     }
   },
   "dependencies": {
-    "@apollo/client": "^3.8.9",
-    "@aptos-labs/aptos-names-connector": "^0.0.2",
-    "@aptos-labs/ts-sdk": "1.28.0",
-    "@aptos-labs/wallet-adapter-mui-design": "^3.0.14",
-    "@aptos-labs/wallet-adapter-react": "^3.7.0",
+    "@apollo/client": "^3.11.8",
+    "@aptos-labs/ts-sdk": "1.33.0",
+    "@aptos-labs/wallet-adapter-mui-design": "3.0.21",
+    "@aptos-labs/wallet-adapter-react": "3.7.7",
     "@bitget-wallet/aptos-wallet-adapter": "^0.1.2",
-    "@blocto/aptos-wallet-adapter-plugin": "^0.2.10",
+    "@blocto/aptos-wallet-adapter-plugin": "^0.2.11",
     "@download/blockies": "^1.0.3",
-    "@emotion/react": "^11.11.3",
-    "@emotion/styled": "^11.11.0",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
     "@martianwallet/aptos-wallet-adapter": "^0.0.5",
     "@msafe/aptos-wallet-adapter": "^1.1.3",
-    "@mui/icons-material": "^5.15.21",
-    "@mui/material": "^5.15.21",
-    "@mui/system": "^5.15.20",
-    "@mui/x-date-pickers": "^7.18.0",
+    "@mui/icons-material": "^6.1.6",
+    "@mui/material": "^6.1.6",
+    "@mui/system": "^6.1.6",
+    "@mui/x-date-pickers": "^7.22.1",
     "@okwallet/aptos-wallet-adapter": "^0.0.7",
     "@pontem/wallet-adapter-plugin": "^0.2.1",
     "@rise-wallet/wallet-adapter": "^0.1.2",
-    "@sentry/react": "^7.119.1",
-    "@sentry/tracing": "^7.119.1",
-    "@tanstack/react-query": "^5.32.1",
+    "@sentry/react": "^8.36.0",
+    "@tanstack/react-query": "^5.59.19",
     "@tp-lab/aptos-wallet-adapter": "^1.0.1",
     "@trustwallet/aptos-wallet-adapter": "^0.1.6",
     "@welldone-studio/aptos-wallet-adapter": "^0.1.5",
     "aptos": "^1.21.0",
     "aptos-node-checker-client": "0.0.5",
-    "chart.js": "^4.4.1",
-    "date-fns": "^3.2.0",
+    "chart.js": "^4.4.6",
+    "date-fns": "^4.1.0",
+    "eslint-plugin-html": "^8.1.2",
     "graphql": "^16.8.1",
     "js-sha3": "^0.9.3",
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
     "pako": "^2.1.0",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-countdown": "^2.3.5",
-    "react-dom": "^18.2.0",
+    "react-countdown": "^2.3.6",
+    "react-dom": "^18.3.1",
     "react-ga4": "^2.1.0",
     "react-gtm-module": "^2.0.11",
-    "react-hook-form": "^7.49.3",
-    "react-intersection-observer": "^9.5.3",
+    "react-hook-form": "^7.53.1",
+    "react-intersection-observer": "^9.13.1",
     "react-json-view": "^1.21.3",
-    "react-loading-skeleton": "^3.3.1",
-    "react-number-format": "^5.3.1",
-    "react-router-dom": "^6.21.2",
+    "react-loading-skeleton": "^3.5.0",
+    "react-number-format": "^5.4.2",
+    "react-router-dom": "^6.27.0",
     "react-simple-maps": "^3.0.0",
-    "react-syntax-highlighter": "^15.5.0",
-    "recharts": "^2.10.4",
+    "react-syntax-highlighter": "^15.6.1",
+    "recharts": "^2.13.3",
     "statsig-react": "^2.1.0"
   },
   "scripts": {
@@ -93,27 +93,27 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@types/lodash": "^4.14.202",
+    "@types/lodash": "^4.17.13",
     "@types/pako": "^2.0.3",
-    "@types/react": "^18.2.47",
+    "@types/react": "^18.3.12",
     "@types/react-addons-perf": "^15.4.2",
-    "@types/react-dom": "^18.2.18",
-    "@types/react-gtm-module": "^2.0.3",
-    "@types/react-simple-maps": "^3.0.4",
-    "@types/react-syntax-highlighter": "^15.5.11",
+    "@types/react-dom": "^18.3.1",
+    "@types/react-gtm-module": "^2.0.4",
+    "@types/react-simple-maps": "^3.0.6",
+    "@types/react-syntax-highlighter": "^15.5.13",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
-    "@vitejs/plugin-react": "^4.3.2",
+    "@vitejs/plugin-react": "^4.3.3",
     "eslint": "^8.56.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.5",
-    "husky": "^9.0.11",
+    "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
-    "prettier": "^3.1.1",
+    "prettier": "^3.3.3",
     "ts-results": "^3.3.0",
-    "typescript": "^5.3.3",
-    "vite": "^5.4.8",
-    "vite-plugin-svgr": "^4.2.0",
-    "vitest": "^2.1.2"
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10",
+    "vite-plugin-svgr": "^4.3.0",
+    "vitest": "^2.1.4"
   }
 }
