@@ -1,14 +1,17 @@
 import {useQuery} from "@tanstack/react-query";
-import {NetworkName} from "../../constants";
+import {knownAddresses, NetworkName} from "../../constants";
 import {useGlobalState} from "../../global-config/GlobalConfig";
 import {
   fetchJsonResponse,
   getLocalStorageWithExpiry,
   setLocalStorageWithExpiry,
+  standardizeAddress,
 } from "../../utils";
 import {ResponseError} from "../client";
 
 const TTL = 60000; // 1 minute
+
+// TODO: Known scam addresses
 
 function getFetchNameUrl(
   network: NetworkName,
@@ -33,11 +36,24 @@ export function useGetNameFromAddress(
   const queryResult = useQuery<string | null, ResponseError>({
     queryKey: ["ANSName", address, shouldCache, state.network_name],
     queryFn: () => {
-      const cachedName = getLocalStorageWithExpiry(address);
+      const standardizedAddress = standardizeAddress(address);
+      const knownName = knownAddresses[standardizedAddress.toLowerCase()];
+      if (knownName) {
+        return knownName;
+      }
+
+      // Change cache key specifically to invalidate all previous cached keys
+      const cachedName = getLocalStorageWithExpiry(`${address}:name`);
       if (cachedName) {
         return cachedName;
       }
-      return genANSName(address, shouldCache, state.network_name, isValidator);
+      // Ensure there's always .apt at the end
+      return genANSName(
+        address,
+        shouldCache,
+        state.network_name,
+        isValidator,
+      ).then((name) => (name ? `${name}.apt` : null));
     },
   });
 
