@@ -216,12 +216,13 @@ export async function getRecentBlocks(
   count: number,
   client: AptosClient,
 ): Promise<Types.Block[]> {
-  const blocks = [];
+  const blockPromises = [];
+  // Don't await here, or they'll be in serial
   for (let i = 0; i < count; i++) {
-    const block = await client.getBlockByHeight(currentBlockHeight - i, false);
-    blocks.push(block);
+    const block = client.getBlockByHeight(currentBlockHeight - i, false);
+    blockPromises.push(block);
   }
-  return blocks;
+  return Promise.all(blockPromises);
 }
 
 export async function getBalance(
