@@ -1,6 +1,5 @@
 import React from "react";
-import {Box} from "@mui/material";
-import Grid from "@mui/material/Unstable_Grid2";
+import {Box, Grid2} from "@mui/material";
 import {grey} from "../../themes/colors/aptosColorPalette";
 import EmptyValue from "./ContentValue/EmptyValue";
 
@@ -19,14 +18,14 @@ export default function ContentRow({
 }: ContentRowProps) {
   return (
     <Box>
-      <Grid
+      <Grid2
         container
         rowSpacing={0.5}
         columnSpacing={4}
         alignItems="start"
         key={i}
       >
-        <Grid xs={12} sm={3}>
+        <Grid2 container size={{xs: 12, sm: 3}}>
           <Box sx={{fontSize: "0.875rem", color: grey[450]}}>
             {title}
             <Box
@@ -40,18 +39,17 @@ export default function ContentRow({
               <Box sx={{display: "inline-block"}}>{tooltip}</Box>
             </Box>
           </Box>
-        </Grid>
-        <Grid
-          xs={12}
-          sm={9}
+        </Grid2>
+        <Grid2
+          size={{xs: 12, sm: 9}}
           sx={{
             fontSize: "0.8rem",
             overflow: "auto",
           }}
         >
           {value ? <Box>{value}</Box> : <EmptyValue />}
-        </Grid>
-      </Grid>
+        </Grid2>
+      </Grid2>
     </Box>
   );
 }
