@@ -1,5 +1,9 @@
 import React from "react";
-import {grey} from "../../../themes/colors/aptosColorPalette";
+import {
+  aptosColor,
+  grey,
+  negativeColor,
+} from "../../../themes/colors/aptosColorPalette";
 import {APTCurrencyValue} from "./CurrencyValue";
 import GasValue from "./GasValue";
 import {Types} from "aptos";
@@ -21,15 +25,16 @@ export default function GasFeeValue({
   netGasCost = false,
   storageRefund = false,
 }: GasFeeValueProps) {
+  const grossGasUnits = BigInt(gasUsed);
+  const netGasWithoutRefund = BigInt(gasUnitPrice) * grossGasUnits;
+
   const feeStatement = transactionData.events?.find(
     (e) => e.type === "0x1::transaction_fee::FeeStatement",
   );
   if (!feeStatement) {
     return (
       <>
-        <APTCurrencyValue
-          amount={(BigInt(gasUnitPrice) * BigInt(gasUsed)).toString()}
-        />
+        <APTCurrencyValue amount={netGasWithoutRefund.toString()} />
         {showGasUsed === true && (
           <span style={{color: grey[450]}}>
             {" ("}
@@ -41,40 +46,45 @@ export default function GasFeeValue({
     );
   }
 
-  const execution_gas = feeStatement.data.execution_gas_units;
-  const io_gas = feeStatement.data.io_gas_units;
-  const storage_gas = feeStatement.data.storage_fee_octas;
-  const storage_refund = feeStatement.data.storage_fee_refund_octas;
+  const feeStatementGasUnits = BigInt(feeStatement.data.total_charge_gas_units);
+  const feeStatementGasUnitsCost = feeStatementGasUnits * BigInt(gasUnitPrice);
+  const storageRefundOctas = BigInt(feeStatement.data.storage_fee_refund_octas);
 
-  const gross_gas_units =
-    BigInt(execution_gas) + BigInt(io_gas) + BigInt(storage_gas);
-  const gross_gas_cost = gross_gas_units * BigInt(gasUnitPrice);
-  const net_gas_cost = gross_gas_cost - BigInt(storage_refund);
+  const netGasWithRefund = feeStatementGasUnitsCost - storageRefundOctas;
 
   if (storageRefund) {
     return (
       <>
-        <APTCurrencyValue amount={storage_refund.toString()} />
+        <APTCurrencyValue amount={storageRefundOctas.toString()} />
       </>
     );
   }
 
   if (netGasCost) {
+    let amountAbs = netGasWithRefund;
+    let color = undefined;
+    if (netGasWithRefund < 0) {
+      color = aptosColor;
+    } else if (netGasWithRefund > 0) {
+      color = negativeColor;
+      amountAbs = -netGasWithRefund;
+    }
     return (
-      <>
-        {net_gas_cost < 0 ? "-" : ""}
-        <APTCurrencyValue amount={net_gas_cost.toString()} />
-      </>
+      <span style={{color: color}}>
+        {netGasWithRefund < 0 && <>+</>}
+        {netGasWithRefund > 0 && <>-</>}
+        <APTCurrencyValue amount={amountAbs.toString()} />
+      </span>
     );
   }
 
   return (
     <>
-      <APTCurrencyValue amount={gross_gas_cost.toString()} />
+      <APTCurrencyValue amount={netGasWithoutRefund.toString()} />
       {showGasUsed === true && (
         <span style={{color: grey[450]}}>
           {" ("}
-          <GasValue gas={gross_gas_units.toString()} />
+          <GasValue gas={grossGasUnits.toString()} />
           {")"}
         </span>
       )}
