@@ -8,20 +8,26 @@ import {
   Typography,
   useTheme,
 } from "@mui/material";
-import {grey} from "../themes/colors/aptosColorPalette";
+import {
+  codeBlockColor,
+  codeBlockColorClickableOnHover,
+  grey,
+} from "../themes/colors/aptosColorPalette";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import {truncateAddress, truncateAddressMiddle} from "../pages/utils";
 import {useGetNameFromAddress} from "../api/hooks/useGetANS";
+import VerifiedOutlined from "@mui/icons-material/VerifiedOutlined";
 
-const BUTTON_HEIGHT = 34;
-const TOOLTIP_TIME = 2000; // 2s
+export const BUTTON_HEIGHT = 34;
+export const TOOLTIP_TIME = 2000; // 2s
 
 export enum HashType {
   ACCOUNT = "account",
   TRANSACTION = "transaction",
   NAME = "name",
   STRUCT = "struct",
+  SYMBOL = "symbol",
 }
 
 interface TitleHashButtonProps {
@@ -120,28 +126,53 @@ function Name({address, isValidator}: {address: string; isValidator: boolean}) {
   if (!name) {
     return null;
   }
+  const isAns = name.endsWith(".apt") || name.endsWith(".petra");
+
+  const ansStyle = {
+    height: BUTTON_HEIGHT,
+    backgroundColor: `${theme.palette.mode === "dark" ? grey[600] : grey[200]}`,
+    borderRadius: 1,
+    color: "inherit",
+    padding: "0.15rem 1rem 0.15rem 1rem",
+  };
+
+  const knownNameStyle = {
+    height: BUTTON_HEIGHT,
+    backgroundColor: codeBlockColor,
+    "&:hover": {
+      backgroundColor: codeBlockColorClickableOnHover,
+    },
+    color: theme.palette.mode === "dark" ? "#83CCED" : "#0EA5E9",
+    padding: "0.15rem 0.35rem 0.15rem 1rem",
+    overflow: "hidden",
+    whiteSpace: "nowrap",
+    textOverflow: "ellipsis",
+    borderRadius: 50,
+    textDecoration: "none",
+  };
 
   return (
     <Box>
-      <Stack
-        justifyContent="center"
-        sx={{
-          height: BUTTON_HEIGHT,
-          backgroundColor: `${
-            theme.palette.mode === "dark" ? grey[600] : grey[200]
-          }`,
-          borderRadius: 1,
-          color: "inherit",
-          padding: "0.15rem 1rem 0.15rem 1rem",
-        }}
-      >
-        <Link
-          href={`https://www.aptosnames.com/name/${name}`}
-          target="_blank"
-          underline="none"
-        >
-          {`${name}.apt`}
-        </Link>
+      <Stack justifyContent="center" sx={isAns ? ansStyle : knownNameStyle}>
+        {isAns ? (
+          <Link
+            href={`https://www.aptosnames.com/name/${name}`}
+            target="_blank"
+            underline="none"
+          >
+            {name}
+          </Link>
+        ) : (
+          <Typography sx={{display: "flex", alignItems: "row", gap: 1}}>
+            <span>{name}</span>
+            <Tooltip title={"This is a verified address label."}>
+              <VerifiedOutlined
+                fontSize="small"
+                sx={{position: "relative", top: 2}}
+              />
+            </Tooltip>
+          </Typography>
+        )}
       </Stack>
     </Box>
   );
