@@ -1,11 +1,29 @@
 import * as React from "react";
-import {Table, TableHead, TableRow} from "@mui/material";
+import {useEffect} from "react";
+import {
+  Button,
+  Stack,
+  Table,
+  TableHead,
+  TableRow,
+  Typography,
+} from "@mui/material";
 import GeneralTableRow from "../../../components/Table/GeneralTableRow";
 import GeneralTableHeaderCell from "../../../components/Table/GeneralTableHeaderCell";
 import HashButton, {HashType} from "../../../components/HashButton";
-import {grey} from "../../../themes/colors/aptosColorPalette";
+import {grey, primary} from "../../../themes/colors/aptosColorPalette";
 import GeneralTableBody from "../../../components/Table/GeneralTableBody";
 import GeneralTableCell from "../../../components/Table/GeneralTableCell";
+import {CoinDescription} from "../../../api/hooks/useGetCoinList";
+import {
+  VerifiedCoinCell,
+  verifiedLevel,
+  VerifiedType,
+} from "../../../components/Table/VerifiedCell";
+import {getAssetSymbol} from "../../../utils";
+import {getLearnMoreTooltip} from "../../Transaction/helpers";
+import {useGlobalState} from "../../../global-config/GlobalConfig";
+import {Network} from "@aptos-labs/ts-sdk";
 
 function CoinNameCell({name}: {name: string}) {
   return (
@@ -44,15 +62,9 @@ function AmountCell({
   );
 }
 
-function CoinTypeCell({
-  assetVersion,
-  assetType,
-}: {
-  assetVersion: string;
-  assetType: string;
-}) {
+function CoinTypeCell({data}: {data: CoinDescriptionPlusAmount}) {
   function getType() {
-    switch (assetVersion) {
+    switch (data.tokenStandard) {
       case "v1":
         return HashType.COIN;
       case "v2":
@@ -64,38 +76,241 @@ function CoinTypeCell({
 
   return (
     <GeneralTableCell sx={{width: 450}}>
-      <HashButton hash={assetType} type={getType()} size="large" />
+      <HashButton
+        hash={data.tokenAddress ?? data.faAddress ?? "Unknown"}
+        type={getType()}
+        size="large"
+        img={data.logoUrl ? data.logoUrl : data.symbol}
+      />
     </GeneralTableCell>
   );
 }
 
-export function CoinsTable({
-  coins,
-}: {
-  coins: {
-    name: string;
-    amount: number;
-    decimals: number;
-    symbol: string;
-    assetType: string;
-    assetVersion: string;
-  }[];
-}) {
+function CoinVerifiedCell({data}: {data: CoinDescriptionPlusAmount}) {
+  return VerifiedCoinCell({
+    data: {
+      id: data.tokenAddress ?? data.faAddress ?? "Unknown",
+      known: data.chainId !== 0,
+      isBanned: data.isBanned,
+      isInPanoraTokenList: data.isInPanoraTokenList,
+      symbol: data?.panoraSymbol ?? data.symbol,
+    },
+  });
+}
+
+enum CoinVerificationFilterType {
+  VERIFIED,
+  RECOGNIZED,
+  ALL,
+  NONE, // Turns it off entirely
+}
+
+export type CoinDescriptionPlusAmount = {
+  amount: number;
+  tokenStandard: string;
+} & CoinDescription;
+
+export function CoinsTable({coins}: {coins: CoinDescriptionPlusAmount[]}) {
+  const [state] = useGlobalState();
+  const [verificationFilter, setVerificationFilter] = React.useState(
+    CoinVerificationFilterType.NONE,
+  );
+
+  useEffect(() => {
+    if (state.network_name === Network.MAINNET) {
+      setVerificationFilter(CoinVerificationFilterType.VERIFIED);
+    }
+  }, [state, state.network_value]);
+
+  function toIndex(coin: CoinDescriptionPlusAmount): number {
+    return coin.panoraOrderIndex
+      ? coin.panoraOrderIndex
+      : coin.chainId !== 0
+        ? 0
+        : 1000000;
+  }
+
+  let filteredCoins = coins;
+
+  function getCoinId(coin: CoinDescriptionPlusAmount): string | null {
+    return coin.tokenAddress ?? coin.faAddress;
+  }
+
+  // TODO: This doesn't cover FAs converted from coins.  The logic for verification has gotten pretty out of hand
+  // and needs to be consolidated before going any further
+  const coinVerifications: Record<string, VerifiedType> = {};
+
+  coins.forEach((coin) => {
+    const coinId = getCoinId(coin);
+    if (coinId) {
+      coinVerifications[coinId] = verifiedLevel(
+        {
+          id: coin.tokenAddress ?? coin.faAddress ?? "Unknown",
+          known: coin.chainId !== 0,
+          isBanned: coin.isBanned,
+          isInPanoraTokenList: coin.isInPanoraTokenList,
+          symbol: coin?.panoraSymbol ?? coin.symbol,
+        },
+        state.network_name,
+      ).level;
+    }
+  });
+
+  switch (verificationFilter) {
+    case CoinVerificationFilterType.VERIFIED:
+      filteredCoins = coins.filter((coin) => {
+        const coinId = getCoinId(coin);
+        if (coinId && coinVerifications[coinId]) {
+          const level = coinVerifications[coinId];
+          return (
+            level === VerifiedType.LABS_VERIFIED ||
+            level === VerifiedType.COMMUNITY_VERIFIED ||
+            level === VerifiedType.NATIVE_TOKEN
+          );
+        } else {
+          return false;
+        }
+      });
+      break;
+    case CoinVerificationFilterType.RECOGNIZED:
+      filteredCoins = coins.filter((coin) => {
+        const coinId = getCoinId(coin);
+        if (coinId && coinVerifications[coinId]) {
+          const level = coinVerifications[coinId];
+          return (
+            level === VerifiedType.LABS_VERIFIED ||
+            level === VerifiedType.COMMUNITY_VERIFIED ||
+            level === VerifiedType.NATIVE_TOKEN ||
+            level === VerifiedType.RECOGNIZED
+          );
+        } else {
+          return false;
+        }
+      });
+
+      break;
+    case CoinVerificationFilterType.ALL:
+    case CoinVerificationFilterType.NONE:
+      filteredCoins = coins;
+      break;
+  }
+  filteredCoins = filteredCoins.sort((a, b) => toIndex(a) - toIndex(b));
+
+  const selectedTextColor = primary[500];
+  const unselectedTextColor = grey[400];
+  const dividerTextColor = grey[200];
+
+  const filterSelector = (
+    <Stack
+      direction="row"
+      justifyContent="flex-end"
+      spacing={1}
+      marginY={0.5}
+      height={16}
+    >
+      <Button
+        variant="text"
+        onClick={() =>
+          setVerificationFilter(CoinVerificationFilterType.VERIFIED)
+        }
+        sx={{
+          fontSize: 12,
+          fontWeight: 600,
+          color:
+            CoinVerificationFilterType.VERIFIED === verificationFilter
+              ? selectedTextColor
+              : unselectedTextColor,
+          padding: 0,
+          "&:hover": {
+            background: "transparent",
+          },
+        }}
+      >
+        Verified
+      </Button>
+      <Typography
+        variant="subtitle1"
+        sx={{
+          fontSize: 11,
+          fontWeight: 600,
+          color: dividerTextColor,
+        }}
+      >
+        |
+      </Typography>
+      <Button
+        variant="text"
+        onClick={() =>
+          setVerificationFilter(CoinVerificationFilterType.RECOGNIZED)
+        }
+        sx={{
+          fontSize: 12,
+          fontWeight: 600,
+          color:
+            CoinVerificationFilterType.RECOGNIZED === verificationFilter
+              ? selectedTextColor
+              : unselectedTextColor,
+          padding: 0,
+          "&:hover": {
+            background: "transparent",
+          },
+        }}
+      >
+        Recognized
+      </Button>
+      <Typography
+        variant="subtitle1"
+        sx={{
+          fontSize: 11,
+          fontWeight: 600,
+          color: dividerTextColor,
+        }}
+      >
+        |
+      </Typography>
+      <Button
+        variant="text"
+        onClick={() => setVerificationFilter(CoinVerificationFilterType.ALL)}
+        sx={{
+          fontSize: 12,
+          fontWeight: 600,
+          color:
+            CoinVerificationFilterType.ALL === verificationFilter
+              ? selectedTextColor
+              : unselectedTextColor,
+          padding: 0,
+          "&:hover": {
+            background: "transparent",
+          },
+        }}
+      >
+        All
+      </Button>
+    </Stack>
+  );
+
+  // TODO: For FA, possibly add store as more info
   return (
-    <Table>
-      <TableHead>
-        <TableRow>
-          <GeneralTableHeaderCell header="Name" />
-          <GeneralTableHeaderCell header="Amount" />
-          <GeneralTableHeaderCell header="Type" />
-          <GeneralTableHeaderCell header="Coin Type" />
-        </TableRow>
-      </TableHead>
-      <GeneralTableBody>
-        {coins.map(
-          ({name, amount, decimals, symbol, assetType, assetVersion}, i) => {
-            let friendlyType = assetVersion;
-            switch (assetVersion) {
+    <>
+      {verificationFilter !== CoinVerificationFilterType.NONE && filterSelector}
+      <Table>
+        <TableHead>
+          <TableRow>
+            <GeneralTableHeaderCell header="Name" />
+            <GeneralTableHeaderCell header="Asset Type" />
+            <GeneralTableHeaderCell header="Asset" />
+            <GeneralTableHeaderCell
+              header="Verified"
+              tooltip={getLearnMoreTooltip("coin_verification")}
+              isTableTooltip={true}
+            />
+            <GeneralTableHeaderCell header="Amount" />
+          </TableRow>
+        </TableHead>
+        <GeneralTableBody>
+          {filteredCoins.map((coinDesc, i) => {
+            let friendlyType = coinDesc.tokenStandard;
+            switch (friendlyType) {
               case "v1":
                 friendlyType = "Coin";
                 break;
@@ -105,22 +320,24 @@ export function CoinsTable({
             }
             return (
               <GeneralTableRow key={i}>
-                <CoinNameCell name={name} />
-                <AmountCell
-                  amount={amount}
-                  decimals={decimals}
-                  symbol={symbol}
-                />
+                <CoinNameCell name={coinDesc.name} />
                 <CoinNameCell name={friendlyType} />
-                <CoinTypeCell
-                  assetVersion={assetVersion}
-                  assetType={assetType}
+                <CoinTypeCell data={coinDesc} />
+                <CoinVerifiedCell data={coinDesc} />
+                <AmountCell
+                  amount={coinDesc.amount}
+                  decimals={coinDesc.decimals}
+                  symbol={getAssetSymbol(
+                    coinDesc.panoraSymbol,
+                    coinDesc.bridge,
+                    coinDesc.symbol,
+                  )}
                 />
               </GeneralTableRow>
             );
-          },
-        )}
-      </GeneralTableBody>
-    </Table>
+          })}
+        </GeneralTableBody>
+      </Table>
+    </>
   );
 }
