@@ -1,4 +1,4 @@
-import {Grid, Box} from "@mui/material";
+import {Grid2, Box} from "@mui/material";
 import React from "react";
 import Typography from "@mui/material/Typography";
 
@@ -11,20 +11,20 @@ type RowProps = {
 export default function Row({title, value, i}: RowProps) {
   return (
     <Box>
-      <Grid
+      <Grid2
         container
         direction={{xs: "column", md: "row"}}
         rowSpacing={1}
         columnSpacing={4}
         key={i}
       >
-        <Grid item md={3}>
+        <Grid2 size={{md: 3}}>
           <Typography variant="subtitle1">{title}</Typography>
-        </Grid>
-        <Grid item md={9} sx={{width: 1, overflowWrap: "break-word"}}>
+        </Grid2>
+        <Grid2 size={{md: 9}} sx={{width: 1, overflowWrap: "break-word"}}>
           <Box sx={{fontSize: 13.5}}>{value}</Box>
-        </Grid>
-      </Grid>
+        </Grid2>
+      </Grid2>
     </Box>
   );
 }
