@@ -1,28 +1,10 @@
 import React from "react";
-import {gql, useQuery} from "@apollo/client";
 import EmptyTabContent from "../../../components/IndividualPageContent/EmptyTabContent";
 import {Types} from "aptos";
-import {CoinsTable} from "../Components/CoinsTable";
-import {normalizeAddress} from "../../../utils";
-
-const COINS_QUERY = gql`
-  query CoinsData($owner_address: String, $limit: Int, $offset: Int) {
-    current_fungible_asset_balances(
-      where: {owner_address: {_eq: $owner_address}}
-      limit: $limit
-      offset: $offset
-    ) {
-      amount
-      asset_type
-      metadata {
-        name
-        decimals
-        symbol
-        token_standard
-      }
-    }
-  }
-`;
+import {CoinDescriptionPlusAmount, CoinsTable} from "../Components/CoinsTable";
+import {useGetCoinList} from "../../../api/hooks/useGetCoinList";
+import {findCoinData} from "../../Transaction/Tabs/BalanceChangeTab";
+import {useGetAccountCoins} from "../../../api/hooks/useGetAccountCoins";
 
 type TokenTabsProps = {
   address: string;
@@ -30,48 +12,70 @@ type TokenTabsProps = {
 };
 
 export default function CoinsTab({address}: TokenTabsProps) {
-  const addr64Hash = normalizeAddress(address);
+  const {data: coinData} = useGetCoinList();
 
-  const {loading, error, data} = useQuery<{
-    current_fungible_asset_balances: {
-      amount: number;
-      asset_type: string;
-      metadata: {
-        name: string;
-        decimals: number;
-        symbol: string;
-        token_standard: string;
-      };
-    }[];
-  }>(COINS_QUERY, {
-    variables: {
-      owner_address: addr64Hash,
-    },
-  });
+  const {isLoading, error, data} = useGetAccountCoins(address);
 
-  if (loading || error) {
-    // TODO: error handling
+  if (isLoading) {
     return null;
   }
 
-  const coins = data?.current_fungible_asset_balances ?? [];
+  if (error) {
+    console.error(error);
+    return null;
+  }
+
+  const coins = data ?? [];
 
   if (coins.length === 0) {
     return <EmptyTabContent />;
   }
 
-  return (
-    <CoinsTable
-      coins={coins
-        .filter((coin) => Boolean(coin.metadata))
-        .map((coin) => ({
-          name: coin.metadata.name,
-          amount: coin.amount,
-          decimals: coin.metadata.decimals,
-          symbol: coin.metadata.symbol,
-          assetType: coin.asset_type,
-          assetVersion: coin.metadata.token_standard,
-        }))}
-    />
-  );
+  function parse_coins(): CoinDescriptionPlusAmount[] {
+    if (!coins || coins.length <= 0) {
+      return [];
+    }
+    return coins
+      .filter((coin) => Boolean(coin.metadata))
+      .map((coin) => {
+        const foundCoin = findCoinData(coinData?.data ?? [], coin.asset_type);
+
+        if (!foundCoin) {
+          // Minimally, return the information we do know
+          return {
+            name: coin.metadata.name,
+            amount: coin.amount,
+            decimals: coin.metadata.decimals,
+            symbol: coin.metadata.symbol,
+            assetType: coin.asset_type,
+            assetVersion: coin.metadata.token_standard,
+            chainId: 0,
+            tokenAddress:
+              coin.metadata.token_standard === "v1" ? coin.asset_type : null,
+            faAddress:
+              coin.metadata.token_standard === "v2" ? coin.asset_type : null,
+            bridge: null,
+            panoraSymbol: null,
+            logoUrl: "",
+            websiteUrl: null,
+            category: "N/A",
+            isInPanoraTokenList: false,
+            isBanned: false,
+            panoraOrderIndex: 20000000,
+            coinGeckoId: null,
+            coinMarketCapId: null,
+            tokenStandard: coin.metadata.token_standard,
+          };
+        } else {
+          // Otherwise, use the stuff found in the lookup
+          return {
+            ...foundCoin,
+            amount: coin.amount,
+            tokenStandard: coin.metadata.token_standard,
+          };
+        }
+      });
+  }
+
+  return <CoinsTable coins={parse_coins()} />;
 }
