@@ -5,16 +5,22 @@ import ContentBox from "../../../components/IndividualPageContent/ContentBox";
 import ContentRow from "../../../components/IndividualPageContent/ContentRow";
 import EmptyTabContent from "../../../components/IndividualPageContent/EmptyTabContent";
 import {getLearnMoreTooltip} from "../../Transaction/helpers";
+import {grey} from "../../../themes/colors/aptosColorPalette";
+import {tryStandardizeAddress} from "../../../utils";
 
 type InfoTabProps = {
   address: string;
   accountData: Types.AccountData | Types.MoveResource[] | undefined;
 };
 
-export default function InfoTab({accountData}: InfoTabProps) {
+export default function InfoTab({address, accountData}: InfoTabProps) {
   if (!accountData || Array.isArray(accountData)) {
     return <EmptyTabContent />;
   }
+
+  const keyRotated =
+    tryStandardizeAddress(address) !==
+    tryStandardizeAddress(accountData.authentication_key);
 
   return (
     <Box marginBottom={3}>
@@ -24,11 +30,24 @@ export default function InfoTab({accountData}: InfoTabProps) {
           value={accountData.sequence_number}
           tooltip={getLearnMoreTooltip("sequence_number")}
         />
-        <ContentRow
-          title={"Authentication Key:"}
-          value={accountData.authentication_key}
-          tooltip={getLearnMoreTooltip("authentication_key")}
-        />
+        {keyRotated ? (
+          <ContentRow
+            title={"Authentication Key:"}
+            value={
+              <>
+                {`${accountData.authentication_key} `}
+                <span style={{marginLeft: 8, color: grey[450]}}>(rotated)</span>
+              </>
+            }
+            tooltip={getLearnMoreTooltip("authentication_key")}
+          />
+        ) : (
+          <ContentRow
+            title={"Authentication Key:"}
+            value={accountData.authentication_key}
+            tooltip={getLearnMoreTooltip("authentication_key")}
+          />
+        )}
       </ContentBox>
     </Box>
   );
