@@ -3,7 +3,7 @@ import {
   TextField,
   Box,
   Divider,
-  Grid,
+  Grid2,
   Stack,
   Typography,
   useMediaQuery,
@@ -30,7 +30,9 @@ import {accountPagePath} from "../../Index";
 interface ModuleSidebarProps {
   sortedPackages: PackageMetadata[];
   selectedModuleName: string;
+
   getLinkToModule(moduleName: string): string;
+
   navigateToModule(moduleName: string): void;
 }
 
@@ -84,16 +86,16 @@ function ViewCode({
   }
 
   return (
-    <Grid container spacing={2}>
-      <Grid item md={3} xs={12}>
+    <Grid2 container spacing={2}>
+      <Grid2 size={{md: 3, xs: 12}}>
         <ModuleSidebar
           sortedPackages={sortedPackages}
           selectedModuleName={selectedModuleName}
           getLinkToModule={getLinkToModule}
           navigateToModule={navigateToModule}
         />
-      </Grid>
-      <Grid item md={9} xs={12}>
+      </Grid2>
+      <Grid2 size={{md: 9, xs: 12}}>
         {selectedModule === undefined ? (
           <EmptyTabContent
             message={`No module found with name: ${selectedModuleName}`}
@@ -105,8 +107,8 @@ function ViewCode({
             bytecode={selectedModule.source}
           />
         )}
-      </Grid>
-    </Grid>
+      </Grid2>
+    </Grid2>
   );
 }
 
