@@ -1,5 +1,5 @@
 import React, {useState} from "react";
-import {Grid} from "@mui/material";
+import {Grid2} from "@mui/material";
 import {useGetAnalyticsData} from "../../api/hooks/useGetAnalyticsData";
 import ChartRangeDaysSelect, {
   ChartRangeDays,
@@ -28,58 +28,58 @@ export default function MainnetAnalytics() {
   }
 
   return (
-    <Grid container spacing={3} marginTop={3}>
-      <Grid item xs={12} md={12} lg={12} marginBottom={2}>
+    <Grid2 container spacing={3} marginTop={3}>
+      <Grid2 size={{xs: 12, md: 12, lg: 12}} marginBottom={2}>
         <NetworkInfo />
-      </Grid>
-      <Grid item xs={12} md={12} lg={12}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 12, lg: 12}}>
         <ChartRangeDaysSelect days={days} setDays={setDays} />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyUserTransactionsChart
           data={data.daily_user_transactions}
           days={days}
         />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyPeakTPSChart data={data.daily_max_tps_15_blocks} days={days} />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <MonthlyActiveUserChart data={data.mau_signers} days={days} />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyActiveUserChart data={data.daily_active_users} days={days} />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyNewAccountsCreatedChart
           data={data.daily_new_accounts_created}
           days={days}
         />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyDeployedContractsChart
           data={data.daily_deployed_contracts}
           days={days}
         />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyContractDeployersChart
           data={data.daily_contract_deployers}
           days={days}
         />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyGasConsumptionChart
           data={data.daily_gas_from_user_transactions}
           days={days}
         />
-      </Grid>
-      <Grid item xs={12} md={6} lg={3}>
+      </Grid2>
+      <Grid2 size={{xs: 12, md: 6, lg: 3}}>
         <DailyAvgGasUnitPriceChart
           data={data.daily_average_gas_unit_price}
           days={days}
         />
-      </Grid>
-    </Grid>
+      </Grid2>
+    </Grid2>
   );
 }
