@@ -1,4 +1,4 @@
-import {Grid, Stack} from "@mui/material";
+import {Grid2, Stack} from "@mui/material";
 import {useParams} from "react-router-dom";
 import PageHeader from "../layout/PageHeader";
 import ValidatorTitle from "./Title";
@@ -20,14 +20,11 @@ import Error from "../Account/Error";
 import {useGetDelegationNodeInfo} from "../../api/hooks/useGetDelegationNodeInfo";
 import {Banner} from "../../components/Banner";
 import {useGetDelegationNodeCommissionChange} from "../../api/hooks/useGetDelegationNodeCommissionChange";
-import {AccountAddress} from "@aptos-labs/ts-sdk";
+import {standardizeAddress} from "../../utils";
 
 export default function ValidatorPage() {
   const address = useParams().address ?? "";
-  const addressHex = useMemo(
-    () => AccountAddress.from(address).toStringLong(),
-    [address],
-  );
+  const addressHex = useMemo(() => standardizeAddress(address), [address]);
   const {validators} = useGetValidators();
   const {connected} = useWallet();
   const {data: accountResource, error} = useGetAccountResource(
@@ -102,9 +99,9 @@ export default function ValidatorPage() {
       }}
     >
       <SkeletonTheme>
-        <Grid container>
+        <Grid2 container>
           <PageHeader />
-          <Grid item xs={12}>
+          <Grid2 size={{xs: 12}}>
             <Stack direction="column" spacing={4}>
               <ValidatorTitle
                 address={address}
@@ -136,8 +133,8 @@ export default function ValidatorPage() {
                 />
               )}
             </Stack>
-          </Grid>
-        </Grid>
+          </Grid2>
+        </Grid2>
       </SkeletonTheme>
     </DelegationStateContext.Provider>
   );
