@@ -15,7 +15,13 @@ import TransactionFunction from "./Components/TransactionFunction";
 import TransactionBlockRow from "./Components/TransactionBlockRow";
 import JsonViewCard from "../../../components/IndividualPageContent/JsonViewCard";
 import {parseExpirationTimestamp} from "../../utils";
-import {TransactionActions} from "./Components/TransactionActions";
+import {grey} from "../../../themes/colors/aptosColorPalette";
+import {LearnMoreTooltip} from "../../../components/IndividualPageContent/LearnMoreTooltip";
+import {
+  CoinDescription,
+  useGetCoinList,
+} from "../../../api/hooks/useGetCoinList";
+import {findCoinData} from "./BalanceChangeTab";
 
 function UserTransferOrInteractionRows({
   transaction,
@@ -23,14 +29,17 @@ function UserTransferOrInteractionRows({
   transaction: Types.Transaction;
 }) {
   const counterparty = getTransactionCounterparty(transaction);
-
-  if (!counterparty) {
-    return null;
+  let smartContractAddress: string | undefined;
+  if (
+    "payload" in transaction &&
+    "function" in transaction.payload &&
+    transaction.payload.function.includes("::")
+  ) {
+    smartContractAddress = transaction.payload.function.split("::")[0];
   }
-
   return (
     <>
-      {counterparty.role === "receiver" && (
+      {counterparty && counterparty.role === "receiver" && (
         <ContentRow
           title="Receiver:"
           value={
@@ -39,11 +48,11 @@ function UserTransferOrInteractionRows({
           tooltip={getLearnMoreTooltip("receiver")}
         />
       )}
-      {counterparty.role === "smartContract" && (
+      {smartContractAddress && (
         <ContentRow
           title="Smart Contract:"
           value={
-            <HashButton hash={counterparty.address} type={HashType.ACCOUNT} />
+            <HashButton hash={smartContractAddress} type={HashType.ACCOUNT} />
           }
           tooltip={getLearnMoreTooltip("smartContract")}
         />
@@ -82,6 +91,69 @@ function TransactionAmountRow({transaction}: {transaction: Types.Transaction}) {
   );
 }
 
+type EventAction = Swap | TokenMint | TokenBurn | ObjectTransfer;
+
+type Swap = {
+  actionType: "swap";
+  dex: string;
+  amountIn: number;
+  amountOut: number;
+  assetIn: string;
+  assetOut: string;
+};
+
+type TokenMint = {
+  actionType: "token mint";
+  collection_address: string;
+  token_address: string;
+};
+type TokenBurn = {
+  actionType: "token burn";
+  previous_owner: string;
+  collection_address: string;
+  token_address: string;
+};
+
+type ObjectTransfer = {
+  actionType: "object transfer";
+  address: string;
+  from: string;
+  to: string;
+};
+
+function TransactionActionsRow({
+  transaction,
+}: {
+  transaction: Types.Transaction;
+}) {
+  const events: Types.Event[] =
+    "events" in transaction ? transaction.events : [];
+  const actions = events.map(getEventAction).filter((a) => a !== undefined);
+
+  const {data: coinData} = useGetCoinList();
+
+  return (
+    <ContentRow
+      title="Actions:"
+      value={actions.map((action, i) => {
+        switch (action.actionType) {
+          case "swap":
+            return swapAction(coinData, action, i);
+          case "token mint":
+            return nftMintAction(action, i);
+          case "token burn":
+            return nftBurnAction(action, i);
+          case "object transfer":
+            return objectTransferAction(action, i);
+        }
+      })}
+      tooltip={
+        <LearnMoreTooltip text="Community-curated interpretations of the transaction." />
+      }
+    />
+  );
+}
+
 type UserTransactionOverviewTabProps = {
   transaction: Types.Transaction;
 };
@@ -90,6 +162,26 @@ export default function UserTransactionOverviewTab({
   transaction,
 }: UserTransactionOverviewTabProps) {
   const transactionData = transaction as Types.Transaction_UserTransaction;
+
+  // TODO: pass into gas fee value to reduce searches
+  const feeStatement = transactionData?.events?.find(
+    (e) => e.type === "0x1::transaction_fee::FeeStatement",
+  );
+  let feePayer: string | undefined;
+  if (
+    transactionData?.signature &&
+    "fee_payer_address" in transactionData.signature
+  ) {
+    feePayer = transactionData.signature.fee_payer_address;
+  }
+
+  let secondarySigners: string[] | undefined;
+  if (
+    transactionData?.signature &&
+    "secondary_signer_addresses" in transactionData.signature
+  ) {
+    secondarySigners = transactionData.signature.secondary_signer_addresses;
+  }
 
   return (
     <Box marginBottom={3}>
@@ -111,9 +203,26 @@ export default function UserTransactionOverviewTab({
           }
           tooltip={getLearnMoreTooltip("sender")}
         />
+        {feePayer && (
+          <ContentRow
+            title="Fee Payer:"
+            value={<HashButton hash={feePayer} type={HashType.ACCOUNT} />}
+            tooltip={getLearnMoreTooltip("fee_payer")}
+          />
+        )}
+        {secondarySigners && secondarySigners.length > 0 && (
+          <ContentRow
+            title="Secondary Signers:"
+            value={secondarySigners.map((address) => (
+              <HashButton hash={address} type={HashType.ACCOUNT} />
+            ))}
+            tooltip={getLearnMoreTooltip("secondary_signers")}
+          />
+        )}
         <UserTransferOrInteractionRows transaction={transactionData} />
         <TransactionFunctionRow transaction={transactionData} />
         <TransactionAmountRow transaction={transactionData} />
+        <TransactionActionsRow transaction={transactionData} />
       </ContentBox>
       <ContentBox>
         <TransactionBlockRow version={transactionData.version} />
@@ -156,35 +265,46 @@ export default function UserTransactionOverviewTab({
           }
           tooltip={getLearnMoreTooltip("gas_fee")}
         />
-        <ContentRow
-          title="Storage Refund:"
-          value={
-            <GasFeeValue
-              gasUsed={transactionData.gas_used}
-              gasUnitPrice={transactionData.gas_unit_price}
-              showGasUsed
-              transactionData={transactionData}
-              storageRefund={true}
+        {(feeStatement?.data?.storage_fee_refund_octas ?? 0) > 0 ? (
+          <>
+            <ContentRow
+              title="Storage Refund:"
+              value={
+                <GasFeeValue
+                  gasUsed={transactionData.gas_used}
+                  gasUnitPrice={transactionData.gas_unit_price}
+                  showGasUsed
+                  transactionData={transactionData}
+                  storageRefund={true}
+                />
+              }
+              tooltip={getLearnMoreTooltip("storage_refund")}
             />
-          }
-          tooltip={getLearnMoreTooltip("storage_refund")}
-        />
-        <ContentRow
-          title="Net Gas Changes:"
-          value={
-            <GasFeeValue
-              gasUsed={transactionData.gas_used}
-              gasUnitPrice={transactionData.gas_unit_price}
-              showGasUsed
-              transactionData={transactionData}
-              netGasCost
+            <ContentRow
+              title="Net Gas Changes:"
+              value={
+                <GasFeeValue
+                  gasUsed={transactionData.gas_used}
+                  gasUnitPrice={transactionData.gas_unit_price}
+                  showGasUsed
+                  transactionData={transactionData}
+                  netGasCost
+                />
+              }
+              tooltip={getLearnMoreTooltip("net_gas_fee")}
             />
-          }
-          tooltip={getLearnMoreTooltip("net_gas_fee")}
-        />
+          </>
+        ) : null}
         <ContentRow
           title="Gas Unit Price:"
-          value={<APTCurrencyValue amount={transactionData.gas_unit_price} />}
+          value={
+            <>
+              <APTCurrencyValue amount={transactionData.gas_unit_price} />{" "}
+              <span style={{color: grey[450]}}>
+                ({transactionData.gas_unit_price} Octas)
+              </span>
+            </>
+          }
           tooltip={getLearnMoreTooltip("gas_unit_price")}
         />
         <ContentRow
@@ -222,7 +342,265 @@ export default function UserTransactionOverviewTab({
           tooltip={getLearnMoreTooltip("accumulator_root_hash")}
         />
       </ContentBox>
-      <TransactionActions transaction={transaction} />
     </Box>
   );
+}
+
+// we define parse<...>Event(event: Types.Event) -> string | undefined
+// and getEventAction will simply go over the list of parse functions and return the first non-undefined result
+function getEventAction(event: Types.Event): EventAction | undefined {
+  const parsers = [
+    parseTokenMintEvent,
+    parseTokenBurnEvent,
+    parseObjectTransferEvent,
+    parseThalaSwapV1Event,
+    parseThalaSwapV2Event,
+  ];
+
+  for (const parse of parsers) {
+    const result = parse(event);
+    if (result !== undefined) {
+      return result;
+    }
+  }
+
+  return undefined;
+}
+
+const swapAction = (
+  coinData: {data: CoinDescription[]} | undefined,
+  action: Swap,
+  i: number,
+) => {
+  const assetInCoin = findCoinData(coinData?.data ?? [], action.assetIn);
+  const assetOutCoin = findCoinData(coinData?.data ?? [], action.assetOut);
+
+  return (
+    <Box
+      key={`action-${i}`}
+      sx={{
+        marginBottom: 1,
+        display: "flex",
+        alignItems: "center",
+        gap: 1,
+      }}
+    >
+      {"🔄 Swapped "}
+      {action.amountIn / Math.pow(10, assetInCoin?.decimals ?? 0)}
+      <HashButton
+        hash={action.assetIn}
+        type={
+          action.assetIn.includes("::")
+            ? HashType.COIN
+            : HashType.FUNGIBLE_ASSET
+        }
+        img={assetInCoin?.logoUrl}
+      />
+      for {action.amountOut / Math.pow(10, assetOutCoin?.decimals ?? 0)}
+      <HashButton
+        hash={action.assetOut}
+        type={
+          action.assetOut.includes("::")
+            ? HashType.COIN
+            : HashType.FUNGIBLE_ASSET
+        }
+        img={assetOutCoin?.logoUrl}
+      />
+      on {action.dex}
+    </Box>
+  );
+};
+
+const nftMintAction = (action: TokenMint, i: number) => {
+  return (
+    <Box
+      key={`action-${i}`}
+      sx={{
+        marginBottom: 1,
+        display: "flex",
+        alignItems: "center",
+        gap: 1,
+      }}
+    >
+      {"🏗️ Minted "}
+      {<HashButton hash={action.token_address} type={HashType.OBJECT} />}
+      {" in collection "}
+      {<HashButton hash={action.collection_address} type={HashType.OBJECT} />}
+    </Box>
+  );
+};
+
+const nftBurnAction = (action: TokenBurn, i: number) => {
+  return (
+    <Box
+      key={`action-${i}`}
+      sx={{
+        marginBottom: 1,
+        display: "flex",
+        alignItems: "center",
+        gap: 1,
+      }}
+    >
+      {"🔥️ Burned "}
+      {<HashButton hash={action.token_address} type={HashType.OBJECT} />}
+      {" in collection "}
+      {<HashButton hash={action.collection_address} type={HashType.OBJECT} />}
+      {" from "}
+      {<HashButton hash={action.previous_owner} type={HashType.ACCOUNT} />}
+    </Box>
+  );
+};
+
+const objectTransferAction = (action: ObjectTransfer, i: number) => {
+  return (
+    <Box
+      key={`action-${i}`}
+      sx={{
+        marginBottom: 1,
+        display: "flex",
+        alignItems: "center",
+        gap: 1,
+      }}
+    >
+      {"⏩ Transferred "}
+      {<HashButton hash={action.address} type={HashType.OBJECT} />} {" from "}
+      {<HashButton hash={action.from} type={HashType.ACCOUNT} />} {" to "}
+      {<HashButton hash={action.to} type={HashType.ACCOUNT} />}
+    </Box>
+  );
+};
+
+function parseTokenMintEvent(event: Types.Event): TokenMint | undefined {
+  if (
+    !event.type.startsWith("0x4::collection::Mint") &&
+    !event.type.startsWith("0x4::collection::MintEvent")
+  ) {
+    return undefined;
+  }
+
+  const data: {
+    collection: string;
+    token: string;
+  } = event.data;
+
+  return {
+    actionType: "token mint",
+    collection_address: data.collection,
+    token_address: data.token,
+  };
+}
+
+function parseTokenBurnEvent(event: Types.Event): TokenBurn | undefined {
+  if (
+    !event.type.startsWith("0x4::collection::Burn") &&
+    !event.type.startsWith("0x4::collection::BurnEvent")
+  ) {
+    return undefined;
+  }
+
+  const data: {
+    collection: string;
+    token: string;
+    previous_owner: string;
+  } = event.data;
+
+  return {
+    actionType: "token burn",
+    collection_address: data.collection,
+    token_address: data.token,
+    previous_owner: data.previous_owner,
+  };
+}
+
+function parseObjectTransferEvent(
+  event: Types.Event,
+): ObjectTransfer | undefined {
+  if (
+    !event.type.startsWith("0x1::object::Transfer") &&
+    !event.type.startsWith("0x1::object::TransferEvent")
+  ) {
+    return undefined;
+  }
+
+  const data: {
+    from: string;
+    to: string;
+    object: string;
+  } = event.data;
+
+  return {
+    actionType: "object transfer",
+    from: data.from,
+    to: data.to,
+    address: data.object,
+  };
+}
+
+function parseThalaSwapV1Event(event: Types.Event): Swap | undefined {
+  if (
+    !(
+      event.type.startsWith(
+        "0x48271d39d0b05bd6efca2278f22277d6fcc375504f9839fd73f74ace240861af::weighted_pool::SwapEvent",
+      ) ||
+      event.type.startsWith(
+        "0x48271d39d0b05bd6efca2278f22277d6fcc375504f9839fd73f74ace240861af::stable_pool::SwapEvent",
+      )
+    )
+  ) {
+    return undefined;
+  }
+
+  const typeArgsMatch = event.type.match(/<(.+)>/);
+  const typeArgs = typeArgsMatch
+    ? typeArgsMatch[1].split(",").map((arg) => arg.trim())
+    : [];
+  const data: {
+    amount_in: string;
+    amount_out: string;
+    idx_in: string;
+    idx_out: string;
+  } = event.data;
+  const amountIn = Number(data.amount_in);
+  const amountOut = Number(data.amount_out);
+  const assetIn = typeArgs[Number(data.idx_in)];
+  const assetOut = typeArgs[Number(data.idx_out)];
+
+  return {
+    actionType: "swap",
+    dex: "ThalaSwap v1",
+    amountIn,
+    amountOut,
+    assetIn,
+    assetOut,
+  };
+}
+
+function parseThalaSwapV2Event(event: Types.Event): Swap | undefined {
+  if (
+    event.type !==
+    "0x7730cd28ee1cdc9e999336cbc430f99e7c44397c0aa77516f6f23a78559bb5::pool::SwapEvent"
+  ) {
+    return undefined;
+  }
+
+  const data: {
+    amount_in: string;
+    amount_out: string;
+    idx_in: string;
+    idx_out: string;
+    metadata: {inner: string}[];
+  } = event.data;
+  const amountIn = Number(data.amount_in);
+  const amountOut = Number(data.amount_out);
+  const assetIn = data.metadata[Number(data.idx_in)].inner;
+  const assetOut = data.metadata[Number(data.idx_out)].inner;
+
+  return {
+    actionType: "swap",
+    dex: "ThalaSwap v2",
+    amountIn,
+    amountOut,
+    assetIn,
+    assetOut,
+  };
 }
